requirements.txt
apache-airflow==2.10.4
boto3>=1.34.0,<2.0.0
pandas>=2.1.0,<3.0.0
numpy>=1.26.0,<2.0.0
pyarrow>=14.0.0,<17.0.0
requests>=2.31.0,<3.0.0
